Authentication and Authorization
Authentication added at base level along with role based and access based authorization to the student and faculty module.

+++ backend/routes/faculty.js
@@ -1,9 +1,14 @@
 const express = require("express");
 const Faculty = require("../models/Faculty");
+const authenticate = require("../middleware/authMiddleware");
+const authorize = require("../middleware/authorize");
 
 const router = express.Router();
 
-router.post("/", async function (req, res) {
+router.post("/",
+  authenticate,
+  authorize("admin"),
+  async function (req, res) {
   try {
     const faculty = await Faculty.create(req.body);
     res.status(201).json({ message: "Faculty created successfully", faculty: faculty });
@@ -13,7 +18,10 @@ router.post("/", async function (req, res) {
   }
 });
 
-router.get("/", async function (req, res) {
+router.get("/",
+  authenticate,
+  authorize("admin", "faculty"),
+  async function (req, res) {
   try {
     const faculty = await Faculty.find();
     res.status(200).json({ message: "Faculty fetched successfully", faculty: faculty });
@@ -23,7 +31,10 @@ router.get("/", async function (req, res) {
   }
 });
 
-router.get("/:facultyId", async function (req, res) {
+router.get("/:facultyId",
+  authenticate,
+  authorize("admin", "faculty"),
+  async function (req, res) {
   try {
     const faculty = await Faculty.findById(req.params.facultyId);
     if (!faculty) return res.status(404).json({ message: "Faculty member not found" });
@@ -34,7 +45,10 @@ router.get("/:facultyId", async function (req, res) {
   }
 });
 
-router.put("/:facultyId", async function (req, res) {
+router.put("/:facultyId",
+  authenticate,
+  authorize("admin"),
+  async function (req, res) {
   try {
     const faculty = await Faculty.findByIdAndUpdate(req.params.facultyId, req.body, { new: true, runValidators: true });
     if (!faculty) return res.status(404).json({ message: "Faculty member not found" });
@@ -45,7 +59,10 @@ router.put("/:facultyId", async function (req, res) {
   }
 });
 
-router.delete("/:facultyId", async function (req, res) {
+router.delete("/:facultyId",
+  authenticate,
+  authorize("admin"),
+  async function (req, res) {
   try {
     const faculty = await Faculty.findByIdAndDelete(req.params.facultyId);
     if (!faculty) return res.status(404).json({ message: "Faculty member not found" });
